refactor: list 조회, profile 생성 facade 분리
